requirements.txt
fastapi
uvicorn
transformers
torch
pillow
python-multipart
streamlit
requests
